fix: report /request cookies exactly as sent
/request built its cookies field from cookie-parser's req.cookies, which
drops a cookie named proto and turns 'j:'-prefixed values into parsed
objects. The endpoint reported something the client never sent, which is
the same gap /cookies had before it started parsing the header itself.

Extract that parsing into src/utils/cookies.ts as parseCookieHeader and
call it from both /request and the cookies service. cookie-parser is left
with no readers, so drop the middleware and the dependency.

The proto half is fixed upstream in cookie v1.0.0, but express 5.2.1
depends on cookie ^0.7.1, so the local parser stays until express moves
to cookie 2.x.
