+++ package.json
@@ -22,14 +22,12 @@
     "verify": "yarn format:check && yarn lint && yarn typecheck"
   },
   "dependencies": {
-    "cookie-parser": "1.4.7",
     "cors": "2.8.6",
     "express": "5.2.1",
     "pino": "10.3.1"
   },
   "devDependencies": {
     "@eslint/js": "10.0.1",
-    "@types/cookie-parser": "1.4.10",
     "@types/cors": "2.8.19",
     "@types/express": "5.0.6",
     "@types/jest": "30.0.0",
