+++ tests/ut/services/cookies.test.ts
@@ -1,30 +1,16 @@
 import { deleteCookies, readCookies, setCookies } from '../../../src/services/cookies.js';
 
-const parseCases: [string, string | undefined, Record<string, string>][] = [
-  ['a single pair', 'flavor=chocolate', { flavor: 'chocolate' }],
-  [
-    'multiple pairs',
-    'flavor=chocolate; session=abc123',
-    { flavor: 'chocolate', session: 'abc123' },
-  ],
-  ['no header', undefined, {}],
-  ['an empty value', 'flavor=', { flavor: '' }],
-  ['a percent-encoded value', 'flavor=chocolate%20chip', { flavor: 'chocolate chip' }],
-  ['a quoted value', 'flavor="chocolate"', { flavor: 'chocolate' }],
-  ['a malformed percent-encoded value', 'flavor=%zz', { flavor: '%zz' }],
-  ["a cookie named '__proto__'", '__proto__=abc', { ['__proto__']: 'abc' }],
-  ["a 'j:'-prefixed value as a string", 'a=j%3A%7B%22x%22%3A1%7D', { a: 'j:{"x":1}' }],
-  [
-    'a repeated name, keeping the first',
-    'flavor=chocolate; flavor=vanilla',
-    { flavor: 'chocolate' },
-  ],
-  ['an entry without a value', 'flavor', {}],
-];
-
 describe('readCookies', () => {
-  it.each(parseCases)('should parse %s', (_description, header, cookies) => {
-    expect(readCookies(header)).toEqual({ status: 200, body: { cookies } });
+  it('should return the cookies carried by the header', () => {
+    const result = readCookies('flavor=chocolate; session=abc123');
+    expect(result).toEqual({
+      status: 200,
+      body: { cookies: { flavor: 'chocolate', session: 'abc123' } },
+    });
+  });
+
+  it('should return an empty object when there is no header', () => {
+    expect(readCookies(undefined)).toEqual({ status: 200, body: { cookies: {} } });
   });
 });
